+++ CHANGELOG.md
@@ -5,9 +5,10 @@ All notable changes to this project will be documented in this file, following t
 
 - Support for multiple states in one `.msvj` file
 - Configurable transitions/animations between states
-- `Component#apply_selection()` now allows defining of an environment around a Component
 - Configurable transparency of representations
-- Add support for additional/custom properties on each node, allowing users to store custom data
+- Add support for additional/custom properties on each node ("vendor-specific properties")
+  - This allows users to store custom data 
+  - Mol* can be instructed to show non-covalent interactions by providing vendor-specific properties
 - Fixes several issues with defined types
 
 ## [v1.0.0] - 2024-04-10